+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/block/RedstoneLanternBlock.kt
@@ -1,11 +1,13 @@
 package dev.aaronhowser.mods.excessive_utilities.block
 
 import net.minecraft.core.BlockPos
+import net.minecraft.core.Direction
 import net.minecraft.world.InteractionResult
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.context.BlockPlaceContext
 import net.minecraft.world.level.BlockGetter
 import net.minecraft.world.level.Level
+import net.minecraft.world.level.SignalGetter
 import net.minecraft.world.level.block.Block
 import net.minecraft.world.level.block.Blocks
 import net.minecraft.world.level.block.state.BlockState
@@ -19,50 +21,37 @@ import net.minecraft.world.phys.shapes.VoxelShape
 
 //TODO: The model for the numbers etc
 class RedstoneLanternBlock : Block(
-	Properties
-		.ofFullCopy(Blocks.STONE)
-		.lightLevel { state ->
-			if (!state.getValue(LIT)) return@lightLevel 0
-			return@lightLevel state.getValue(POWER)
-		}
+	Properties.ofFullCopy(Blocks.STONE)
 ) {
 
 	init {
 		registerDefaultState(
 			stateDefinition.any()
 				.setValue(POWER, 0)
-				.setValue(LIT, false)
 		)
 	}
 
 	override fun createBlockStateDefinition(builder: StateDefinition.Builder<Block, BlockState>) {
-		builder.add(POWER, LIT)
-	}
-
-	override fun getStateForPlacement(context: BlockPlaceContext): BlockState? {
-		return defaultBlockState()
-			.setValue(LIT, context.level.hasNeighborSignal(context.clickedPos))
+		builder.add(POWER)
 	}
 
 	override fun useWithoutItem(state: BlockState, level: Level, pos: BlockPos, player: Player, hitResult: BlockHitResult): InteractionResult {
 		val oldPower = state.getValue(POWER)
 		val newPower = if (player.isSecondaryUseActive) {
-			if (oldPower == 0) 15 else oldPower - 1
+			if (oldPower <= 0) 15 else oldPower - 1
 		} else {
-			if (oldPower == 15) 0 else oldPower + 1
+			if (oldPower >= 15) 0 else oldPower + 1
 		}
 
 		level.setBlockAndUpdate(pos, state.setValue(POWER, newPower))
 		return InteractionResult.SUCCESS
 	}
 
-	override fun neighborChanged(state: BlockState, level: Level, pos: BlockPos, neighborBlock: Block, neighborPos: BlockPos, movedByPiston: Boolean) {
-		val wasLit = state.getValue(LIT)
-		val shouldBeLit = level.hasNeighborSignal(pos)
+	override fun isSignalSource(state: BlockState): Boolean = true
+	override fun shouldCheckWeakPower(state: BlockState, level: SignalGetter, pos: BlockPos, side: Direction): Boolean = false
 
-		if (wasLit != shouldBeLit) {
-			level.setBlockAndUpdate(pos, state.setValue(LIT, shouldBeLit))
-		}
+	override fun getSignal(state: BlockState, level: BlockGetter, pos: BlockPos, direction: Direction): Int {
+		return state.getValue(POWER)
 	}
 
 	override fun getShape(state: BlockState, level: BlockGetter, pos: BlockPos, context: CollisionContext): VoxelShape {
@@ -71,7 +60,6 @@ class RedstoneLanternBlock : Block(
 
 	companion object {
 		val POWER: IntegerProperty = BlockStateProperties.POWER
-		val LIT: BooleanProperty = BlockStateProperties.LIT
 		val SHAPE: VoxelShape = box(1.0, 1.0, 1.0, 15.0, 15.0, 15.0)
 	}